requirements.txt
numpy>=1.21.0
torch>=1.10.0
pyyaml>=5.4.1
gymnasium>=0.28.1
matplotlib>=3.5.0
tensorboard>=2.8.0
pandas>=1.3.0
streamlit>=1.10.0
altair>=4.0.0
